requirements.txt
matplotlib==3.7.5
pandas==2.0.3
python-dotenv==1.0.1
vaderSentiment==3.3.2

argparse~=1.4.0
beautifulsoup4~=4.12.3